LPS-56048 Move SocialActivity classes into module
