+++ modules/apps/message-boards/message-boards-test/test/integration/com/liferay/message/boards/social/test/MBMessageActivityInterpreterTest.java
@@ -12,24 +12,31 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.social;
+package com.liferay.message.boards.social.test;
 
+import com.liferay.message.boards.web.social.MBMessageActivityInterpreter;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
+import com.liferay.portal.kernel.util.ObjectValuePair;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
 import com.liferay.portlet.messageboards.model.MBCategoryConstants;
 import com.liferay.portlet.messageboards.model.MBMessage;
+import com.liferay.portlet.messageboards.model.MBMessageConstants;
 import com.liferay.portlet.messageboards.service.MBMessageLocalServiceUtil;
 import com.liferay.portlet.messageboards.service.MBThreadLocalServiceUtil;
-import com.liferay.portlet.social.model.SocialActivityConstants;
+import com.liferay.portlet.messageboards.social.MBActivityKeys;
 import com.liferay.portlet.social.model.SocialActivityInterpreter;
 import com.liferay.portlet.social.test.BaseSocialActivityInterpreterTestCase;
+
+import java.io.InputStream;
+
+import java.util.Collections;
 
 import org.junit.ClassRule;
 import org.junit.Rule;
@@ -38,7 +45,7 @@ import org.junit.Rule;
  * @author Zsolt Berentey
  */
 @Sync
-public class MBThreadActivityInterpreterTest
+public class MBMessageActivityInterpreterTest
 	extends BaseSocialActivityInterpreterTestCase {
 
 	@ClassRule
@@ -50,29 +57,44 @@ public class MBThreadActivityInterpreterTest
 
 	@Override
 	protected void addActivities() throws Exception {
+		message = addMessage(null);
+
+		message = addMessage(message);
+	}
+
+	protected MBMessage addMessage(MBMessage parentMessage) throws Exception {
 		ServiceContext serviceContext =
 			ServiceContextTestUtil.getServiceContext(
 				group.getGroupId(), TestPropsValues.getUserId());
 
-		MBMessage message = MBMessageLocalServiceUtil.addMessage(
-			TestPropsValues.getUserId(), RandomTestUtil.randomString(),
-			group.getGroupId(), MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID,
-			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
-			serviceContext);
+		long categoryId = MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID;
+		long parentMessageId = MBMessageConstants.DEFAULT_PARENT_MESSAGE_ID;
+		long threadId = 0;
 
-		_threadId = message.getThreadId();
+		if (parentMessage != null) {
+			categoryId = parentMessage.getCategoryId();
+			parentMessageId = parentMessage.getMessageId();
+			threadId = parentMessage.getThreadId();
+		}
+
+		return MBMessageLocalServiceUtil.addMessage(
+			TestPropsValues.getUserId(), RandomTestUtil.randomString(),
+			group.getGroupId(), categoryId, threadId, parentMessageId,
+			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
+			MBMessageConstants.DEFAULT_FORMAT,
+			Collections.<ObjectValuePair<String, InputStream>>emptyList(),
+			false, 0.0, false, serviceContext);
 	}
 
 	@Override
 	protected SocialActivityInterpreter getActivityInterpreter() {
-		return new MBThreadActivityInterpreter();
+		return new MBMessageActivityInterpreter();
 	}
 
 	@Override
 	protected int[] getActivityTypes() {
 		return new int[] {
-			SocialActivityConstants.TYPE_MOVE_TO_TRASH,
-			SocialActivityConstants.TYPE_RESTORE_FROM_TRASH
+			MBActivityKeys.ADD_MESSAGE, MBActivityKeys.REPLY_MESSAGE
 		};
 	}
 
@@ -83,8 +105,8 @@ public class MBThreadActivityInterpreterTest
 
 	@Override
 	protected void moveModelsToTrash() throws Exception {
-		MBThreadLocalServiceUtil.moveThreadsToTrash(
-			group.getGroupId(), TestPropsValues.getUserId());
+		MBThreadLocalServiceUtil.moveThreadToTrash(
+			TestPropsValues.getUserId(), message.getThreadId());
 	}
 
 	@Override
@@ -94,9 +116,9 @@ public class MBThreadActivityInterpreterTest
 	@Override
 	protected void restoreModelsFromTrash() throws Exception {
 		MBThreadLocalServiceUtil.restoreThreadFromTrash(
-			TestPropsValues.getUserId(), _threadId);
+			TestPropsValues.getUserId(), message.getThreadId());
 	}
 
-	private long _threadId;
+	protected MBMessage message;
 
 }
