+++ modules/apps/message-boards/message-boards-test/test/integration/com/liferay/message/boards/social/test/MBThreadActivityInterpreterTest.java
@@ -12,29 +12,25 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.social;
+package com.liferay.message.boards.social.test;
 
+import com.liferay.message.boards.web.social.MBThreadActivityInterpreter;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.util.ObjectValuePair;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
 import com.liferay.portlet.messageboards.model.MBCategoryConstants;
 import com.liferay.portlet.messageboards.model.MBMessage;
-import com.liferay.portlet.messageboards.model.MBMessageConstants;
 import com.liferay.portlet.messageboards.service.MBMessageLocalServiceUtil;
 import com.liferay.portlet.messageboards.service.MBThreadLocalServiceUtil;
+import com.liferay.portlet.social.model.SocialActivityConstants;
 import com.liferay.portlet.social.model.SocialActivityInterpreter;
 import com.liferay.portlet.social.test.BaseSocialActivityInterpreterTestCase;
-
-import java.io.InputStream;
-
-import java.util.Collections;
 
 import org.junit.ClassRule;
 import org.junit.Rule;
@@ -43,7 +39,7 @@ import org.junit.Rule;
  * @author Zsolt Berentey
  */
 @Sync
-public class MBMessageActivityInterpreterTest
+public class MBThreadActivityInterpreterTest
 	extends BaseSocialActivityInterpreterTestCase {
 
 	@ClassRule
@@ -55,44 +51,29 @@ public class MBMessageActivityInterpreterTest
 
 	@Override
 	protected void addActivities() throws Exception {
-		message = addMessage(null);
-
-		message = addMessage(message);
-	}
-
-	protected MBMessage addMessage(MBMessage parentMessage) throws Exception {
 		ServiceContext serviceContext =
 			ServiceContextTestUtil.getServiceContext(
 				group.getGroupId(), TestPropsValues.getUserId());
 
-		long categoryId = MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID;
-		long parentMessageId = MBMessageConstants.DEFAULT_PARENT_MESSAGE_ID;
-		long threadId = 0;
-
-		if (parentMessage != null) {
-			categoryId = parentMessage.getCategoryId();
-			parentMessageId = parentMessage.getMessageId();
-			threadId = parentMessage.getThreadId();
-		}
-
-		return MBMessageLocalServiceUtil.addMessage(
+		MBMessage message = MBMessageLocalServiceUtil.addMessage(
 			TestPropsValues.getUserId(), RandomTestUtil.randomString(),
-			group.getGroupId(), categoryId, threadId, parentMessageId,
+			group.getGroupId(), MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID,
 			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
-			MBMessageConstants.DEFAULT_FORMAT,
-			Collections.<ObjectValuePair<String, InputStream>>emptyList(),
-			false, 0.0, false, serviceContext);
+			serviceContext);
+
+		_threadId = message.getThreadId();
 	}
 
 	@Override
 	protected SocialActivityInterpreter getActivityInterpreter() {
-		return new MBMessageActivityInterpreter();
+		return new MBThreadActivityInterpreter();
 	}
 
 	@Override
 	protected int[] getActivityTypes() {
 		return new int[] {
-			MBActivityKeys.ADD_MESSAGE, MBActivityKeys.REPLY_MESSAGE
+			SocialActivityConstants.TYPE_MOVE_TO_TRASH,
+			SocialActivityConstants.TYPE_RESTORE_FROM_TRASH
 		};
 	}
 
@@ -103,8 +84,8 @@ public class MBMessageActivityInterpreterTest
 
 	@Override
 	protected void moveModelsToTrash() throws Exception {
-		MBThreadLocalServiceUtil.moveThreadToTrash(
-			TestPropsValues.getUserId(), message.getThreadId());
+		MBThreadLocalServiceUtil.moveThreadsToTrash(
+			group.getGroupId(), TestPropsValues.getUserId());
 	}
 
 	@Override
@@ -114,9 +95,9 @@ public class MBMessageActivityInterpreterTest
 	@Override
 	protected void restoreModelsFromTrash() throws Exception {
 		MBThreadLocalServiceUtil.restoreThreadFromTrash(
-			TestPropsValues.getUserId(), message.getThreadId());
+			TestPropsValues.getUserId(), _threadId);
 	}
 
-	protected MBMessage message;
+	private long _threadId;
 
 }
